Add example/test programs for Erostek ET232
et232-info.py - dump a bunch of interesting registers that demonstrates that you can interact with the device.
et232-cycle-test.py - a  stress test that continually changes the mode/MA/levels to ensures that the device stays accessible and will run for as long as desired.  We discovered it was critical to periodically resent the Auto-Timeout register at 0xd3.